// src/app/dashboard/modules/inbox/modules/message-list-view/interfaces/message.interface.ts
export interface IMessageQueryParams {
  search: string;
  type: EMessageType;
  assignedTo: string[];
  propertyManagerId: string[];
  messageStatus: string[];
  taskDeliveryFailIds: string[];
  topicId: string;
  propertyId: string;
  excludeUnHappyPath: boolean;
  excludeConversation: boolean;
  limit: number;
  // same as limit
  // TODO: verify to remove either one
  pageLimit?: number;
  page: number;
  onlyTask: boolean;
  status: TaskStatusType;
  mailBoxId: string;
  currentTaskId?: string;
  isLoading?: boolean;
  labelId?: string;
  reminderType?: string;
  isShowIgnore?: boolean;
  reminderTime?: number;
  timestamp?: number;
  showMessageInTask?: boolean;
}

export interface IMarkAsUnreadResponse {
  id: string;
  isRead: boolean;
}

export interface IFlagUrgentMessageResponse {
  id: string;
  isUrgent: boolean;
}

export enum EMessageType {
  MY_MESSAGES = 'MY_MESSAGES',
  TEAM_MESSAGES = 'TEAM_MESSAGES',
  UNASSIGNED = 'UNASSIGNED'
}

export enum EMessageQueryType {
  INBOX_TYPE = 'inboxType',
  MESSAGE_STATUS = 'status',
  EXTERNAL_ID = 'externalId',
  MAILBOX_ID = 'mailBoxId',
  THREAD_ID = 'threadId',
  EMAIL_MESSAGE_ID = 'emailMessageId',
  TIME_STAMP = 'timestamp',
  SHOW_MESSAGE_IN_TASK = 'showMessageInTask'
}
export enum EInboxFilterSelected {
  ASSIGNED_TO = 'assignedTo',
  PROPERTY_MANAGER_ID = 'propertyManagerId',
  SEARCH = 'search',
  AI_ASSISTANT_TYPE = 'aiAssistantType',
  CONVERSATION_LOG_ID = 'conversationLogId',
  CONVERSATION_ID = 'conversationId',
  MESSAGE_STATUS = 'messageStatus',
  TASK_ID = 'taskId',
  EXTERNAL_ID = 'externalId',
  TASK_EDITOR_ID = 'taskEditorId',
  SHOW_MESSAGE_IN_TASK = 'showMessageInTask'
}